# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  apperrors.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Stop:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.StopUpdate:
    properties:
      address:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
    type: object
...
  model.Trip:
    properties:
      base_price:
        type: integer
      bus_id:
        type: integer
      created_at:
        type: string
      driver_id:
        type: integer
      end_time:
        type: string
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    required:
    - base_price
    - bus_id
    - driver_id
    - end_time
    - start_time
    type: object
  model.TripShortInfo:
    properties:
      base_price:
        type: number
      brand:
        type: string
      driver_name:
        type: string
      end_stop:
        type: string
      end_time:
        type: string
      license_plate:
        type: string
      start_stop:
        type: string
      start_time:
        type: string
      trip_id:
        type: integer
    type: object
  model.TripStop:
    properties:
      arrival_time:
        type: string
      departure_time:
        type: string
      id:
        type: integer
      price_to_next:
        type: integer
      stop_id:
        type: integer
      stop_order:
        type: integer
      trip_id:
        type: integer
    type: object
...
  model.TripUpdate:
    properties:
      base_price:
        type: integer
      bus_id:
        type: integer
      created_at:
        type: string
      driver_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: luchitstimofei@yandex.ru
    name: Luchits Timofei
    url: http://example.com
  description: This is a server for order private bus.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ordering Bus API
  version: "1.0"
paths:
...
  /admin/stops:
    post:
      consumes:
      - application/json
      description: Создает новую запись остановки (требуются права администратора)
      parameters:
      - description: Данные остановки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TripStop'
      produces:
      - application/json
      responses:
        "201":
          description: Остановка успешно создана
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Создать новую остановки (только админ)
      tags:
      - admin/stops
  /admin/stops/{id}:
    delete:
      description: Удаляет запись об остановке по ID (требуются права администратора)
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Остановка успешно удалена
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Остановка не найдена
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Удалить остановки (только админ)
      tags:
      - admin/stops
    put:
      consumes:
      - application/json
      description: Обновляет информацию об остановке по ID (требуются права администратора)
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные остановки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.StopUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Данные остановки обновлены
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Обновить данные об остановке (только админ)
      tags:
      - admin/stops
...
  /admin/trips:
    post:
      consumes:
      - application/json
      description: Создает новую запись о маршруте (требуются права администратора)
      parameters:
      - description: Данные маршрута
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Trip'
      produces:
      - application/json
      responses:
        "201":
          description: Маршрут успешно создан
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Создать новый маршрут (только админ)
      tags:
      - admin/trips
  /admin/trips/{id}:
    delete:
      description: Удаляет запись о маршруте по ID (требуются права администратора)
      parameters:
      - description: ID маршрута
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Маршрут успешно удален
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Маршрут не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Удалить маршрут (только админ)
      tags:
      - admin/trips
    put:
      consumes:
      - application/json
      description: Обновляет информацию о маршруте по ID (требуются права администратора)
      parameters:
      - description: ID маршрута
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные маршрута
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TripUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Данные маршрута обновлены
          schema:
            $ref: '#/definitions/apperrors.SuccessResponse'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Водитель не найден
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - Bearer: []
      summary: Обновить данные о маршруте (только админ)
      tags:
      - admin/trips
...
  /stops:
    get:
      description: Возвращает все остановки в системе
      produces:
      - application/json
      responses:
        "200":
          description: Модель остановок
          schema:
            $ref: '#/definitions/model.Stop'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить все остановки
      tags:
      - stops
  /stops/{id}:
    get:
      description: Возвращает остановку по ID
      parameters:
      - description: ID остановки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Модель пути
          schema:
            $ref: '#/definitions/model.TripStop'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить остановку
      tags:
      - stops
...
  /trips:
    get:
      description: Возвращает онимичную модель для отображения пользователю информации
        о маршрутах и их остановках
      produces:
      - application/json
      responses:
        "200":
          description: Модель остановок
          schema:
            $ref: '#/definitions/model.TripShortInfo'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить информацию по таблице trip
      tags:
      - trips
  /trips/{id}:
    get:
      description: Возвращает маршрут в системе по ID
      parameters:
      - description: ID маршрута
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Модель пути
          schema:
            $ref: '#/definitions/model.Trip'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Получить маршрут
      tags:
      - trips
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"